# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Onlus website"
  version: "1.0.0"
  title: "AINA onlus"
  contact:
    email: "10641118@polimi.it"
  license:
    name: "AFR"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:4000"
basePath: "/v1"
tags:
- name: "service"
  description: "service provided by the association"
- name: "event"
  description: "event organized by the association"
- name: "person"
  description: "person who work for the association"
- name: "email"
  description: "newsletter"
- name: "rel"
  description: "relationship between people and service"
- name: "users"
  description: "users"
paths:
  /services:
    get:
      tags:
      - "service"
      summary: "get info about a service"
      operationId: "servicesGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "maximum number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "search"
        in: "query"
        description: "generic text search"
        required: false
        type: "string"
      responses:
        "200":
          description: "A collection of services"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "404":
          description: "Unexpected error"
      x-swagger-router-controller: "Service"
  /services/{serviceId}:
    get:
      tags:
      - "service"
      summary: "Find service by ID"
      description: "return a service"
      operationId: "servicesServiceIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "ID of service to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Service"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /events:
    get:
      tags:
      - "event"
      summary: "get info about an event"
      operationId: "eventsGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "maximum number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "search"
        in: "query"
        description: "generic text search"
        required: false
        type: "string"
      responses:
        "200":
          description: "A collection of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        "404":
          description: "Unexpected error"
      x-swagger-router-controller: "Event"
  /events/{eventId}:
    get:
      tags:
      - "event"
      summary: "Find event by ID"
      description: "return an event"
      operationId: "eventsEventIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /people:
    get:
      tags:
      - "person"
      summary: "get info about a person"
      operationId: "peopleGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "maximum number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "search"
        in: "query"
        description: "generic text search"
        required: false
        type: "string"
      responses:
        "200":
          description: "A collection of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Person"
        "404":
          description: "Unexpected error"
      x-swagger-router-controller: "Person"
  /people/{personId}:
    get:
      tags:
      - "person"
      summary: "Find person by ID"
      description: "return a person"
      operationId: "peoplePersonIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "personId"
        in: "path"
        description: "ID of person to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Person"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
  /rel:
    get:
      tags:
      - "rel"
      summary: "get info about the relationship between people and services"
      operationId: "relGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Relationship between people and services"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Rel"
        "404":
          description: "Unexpected error"
      x-swagger-router-controller: "Rel"
  /email:
    post:
      tags:
      - "email"
      summary: "email of the newsletter"
      operationId: "emailPOST"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "newsletter's email"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Email"
        "404":
          description: "Unexpected error"
      x-swagger-router-controller: "Email"
  /users/login:
    post:
      tags:
        - user
      summary: Login
      description: Login with a form
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
      responses:
        '200':
          description: successful login
        '401':
          description: unauthorized
  /users/me:
    get:
      tags:
        - user
      summary: Data of the user of the current session
      produces:
        - application/json
      responses:
        '200':
          description: successful login
          schema:
            $ref: '#/definitions/User'
        '401':
          description: unauthorized

definitions:
  Service:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      photo_url:
        type: "string"
      description:
        type: "string"
    title: "Service"
    description: "service provided by the association"
    example:
      name: "name"
      description: "description"
      id: 0
      photo_url: "photo_url"
  Event:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
      photo_url:
        type: "string"
      description:
        type: "string"
      location:
        type: "string"
      person_id:
        type: "integer"
        format: "int64"
      service_id:
        type: "integer"
        format: "int64"
    title: "Event"
    description: "event provided by the association"
    example:
      date: "date"
      service_id: 1
      name: "name"
      description: "description"
      location: "location"
      id: 0
      photo_url: "photo_url"
      person_id: 6
  Person:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
      photo_url:
        type: "string"
      phone:
        type: "string"
      role:
        type: "string"
      description:
        type: "string"
      main_service:
        type: "integer"
        format: "int64"
    title: "Person"
    description: "person who works for the association"
    example:
      role: "role"
      phone: "phone"
      surname: "surname"
      name: "name"
      description: "description"
      main_service: 6
      id: 0
      photo_url: "photo_url"
      email: "email"
  Rel:
    type: "object"
    properties:
      person_id:
        type: "integer"
        format: "int64"
      service_id:
        type: "integer"
        format: "int64"
    title: "Rel"
    description: "relationship between people and services"
    example:
      service_id: 6
      person_id: 0
  Email:
    type: "object"
    properties:
      email:
        type: "string"
    title: "Email"
    description: "email registered to the newsletter"
    example:
      email: "email"
  User:
    title: User
    description: A user
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      address:
        type: string
      creditcard:
        type: string
    example:
      id: 1
      name: Vic
      address: Via le mani dal naso
      creditcard: asdfgh